+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
   - Actions to tree entries
     - Stash: Apply/pop and Drop
+    - File: Diff to file in current state
 ### Updated
   - Icons are now consistent with the general UI
   - Improve default stash tooltips for tree items